# Conflicts:
#	README.md
#	src/seqtrainer/benchmarks/__init__.py
#	src/seqtrainer/benchmarks/compare.py
#	src/seqtrainer/benchmarks/runner.py
#	src/seqtrainer/benchmarks/splits.py
#	src/seqtrainer/cli/main.py
#	src/seqtrainer/metrics/__init__.py
#	src/seqtrainer/metrics/classification.py
#	src/seqtrainer/torch/cnn_baseline.py
#	tests/test_benchmark_config.py
#	tests/test_benchmark_harness.py
#	tests/test_cnn_baseline.py

+++ src/seqtrainer/benchmarks/splits.py
@@ -53,7 +53,7 @@ def summarize_split_frames(config: BenchmarkConfig, frames: dict[str, pd.DataFra
     summary: dict[str, Any] = {}
     for split in REQUIRED_SPLITS:
         frame = frames[split]
-        labels = frame[config.dataset.label_field].astype(int)
+        labels = frame[config.dataset.label_field]
         class_counts = labels.value_counts().sort_index().to_dict()
         split_summary: dict[str, Any] = {
             "rows": int(len(frame)),
